+++ src/main/java/com/chargebee/exceptions/BatchAPIException.java
@@ -2,6 +2,8 @@ package com.chargebee.exceptions;
 
 import com.chargebee.APIException;
 import org.json.JSONObject;
+import java.util.List;
+import java.util.Map;
 
 import static com.chargebee.BatchConstants.CORRELATION_ID;
 
@@ -12,7 +14,10 @@ public class BatchAPIException extends APIException {
         super(httpStatusCode, message, jsonObj);
         this.correlationId = jsonObj.optString(CORRELATION_ID);
     }
-
+    public BatchAPIException(int httpStatusCode, String message, JSONObject jsonObj, Map<String, List<String>> responseHeaders) {
+        super(httpStatusCode, message, jsonObj, responseHeaders);
+        this.correlationId = jsonObj.optString(CORRELATION_ID);
+    }
     public String getCorrelationId() {
         return correlationId;
     }