+++ CHANGELOG.md
@@ -1,3 +1,16 @@
+### v3.30.0 (2025-03-06)
+* * * 
+
+### New Input Params:
+* created_at has been added to Invoice#LineItemsImportInputParams.
+* PaymentIntent has been added to Purchase#CreateInputParams.
+
+### New Enums:
+* SUBSCRIPTION_ENTITLEMENTS_UPDATED has been added to EventType#Enum.
+
+### New Attribute: 
+* Added response headers in APIException
+
 ### v3.29.0 (2025-02-10)
 * * *